+++ src/intranet3/lib/scrum/__init__.py
@@ -37,7 +37,11 @@ class BugUglyAdapter(object):
         return (points / self.sprint_time) if self.sprint_time else 0.0
 
     @classmethod
-    def produce(cls, bugs):
+    def produce(cls, bugs, project):
+        for bug in bugs:
+            if not bug.project_id:
+                bug.project = project
+                bug.project_id = project.id
         bugs = [BugUglyAdapter(bug) for bug in bugs]
         bugs_dict = dict([(bug.id, bug) for bug in bugs])
 
@@ -82,7 +86,7 @@ def move_blocked_to_the_end(bugs):
 class SprintWrapper(object):
     def __init__(self, sprint, bugs, request):
         self.sprint = sprint
-        self.bugs = BugUglyAdapter.produce(bugs)
+        self.bugs = BugUglyAdapter.produce(bugs, sprint.project)
         self.request = request
         self.session = request.db_session